Moving implementation of Callbacks next to ExecutionService
